// devrush-ctms/src/lib/schema.ts
export type UserRole =
  | "PI"
  | "SUB_INVESTIGATOR"
  | "COORDINATOR"
  | "DATA_MANAGER"
  | "MONITOR"
  | "PV_OFFICER"
  | "EC_MEMBER"
  | "DSMB_MEMBER"
  | "REGULATORY"
  | "SPONSOR"
  | "ADMIN";